Fix connection reset error
When the HTTP server was being run locally and an HTTP GET request was
made with either curl, the browser, or Postman, it sometimes threw an
error stating that the connection was 'reset by peer'.

Now that the HTTP response has been corrected, the connection remains open
after the client receives the response.