Change Header parsing based on acceptance test feedback
The HTTPSpec acceptance test suite and manual acceptance testing
showed that the end of the header section is marked when the
buffered reader reads in an empty string.
In unit testing, when an additional CRLF is sent to the input stream
after the headers section, according to the HTTP/1.1 RFC standard,
the buffered reader reads it as a CRLF.
The additionalHeadersRemain condition was changed to account for
both scenarios.