Fix bug in BodyParser
When a POST request was made, the server continued waiting for
more content to be sent over the input stream even after the
content length was reached. When the client manually stopped the
request, the BodyParser was able to read the body accurately because
the stream was closed.

Now, the BodyParser is able to identify the end of the input stream
and the server no longer hangs up while parsing the request body.